Rename domain module to views_data
